feat(ErrorManager): update to TypeScript
